cli/pyproject.toml
[project]
name = "aletheia-cli"
version = "0.1.0"
description = "Command-line interface for Aletheia document parser"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
dependencies = [
    "click>=8.1.0",
    "httpx>=0.26.0",
    "rich>=13.7.0",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
]

[project.scripts]
aletheia = "aletheia_cli.main:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["aletheia_cli"]
